+++ examples/preact/app.ts
@@ -1,16 +1,6 @@
 import * as Peko from "https://deno.land/x/peko/mod.ts"
-import { HandlerParams } from "https://deno.land/x/peko/lib/types.ts"
-
 import { lookup } from "https://deno.land/x/media_types@v3.0.3/mod.ts"
 import { recursiveReaddir } from "https://deno.land/x/recursive_readdir@v2.0.0/mod.ts"
-import { renderToString } from "https://npm.reversehttp.com/preact,preact/hooks,htm/preact,preact-render-to-string"
-
-// Preact page component imports
-// These are necessary for Deno Deploy support
-import Home from "./src/pages/Home.js"
-import About from "./src/pages/About.js"
-
-import htmlTemplate from "../template.ts"
 import config from "../config.ts"
 
 // Configure Peko
@@ -19,47 +9,7 @@ Peko.setConfig(config)
 console.log(Peko.getConfig())
 
 // SSR'ed app page routes
-const ssrRoutes = [
-    // must be SSRRoute type (see types.ts)
-    {
-        route: "/",
-        // srcURL used for emitting file change events in devMode
-        srcURL: new URL("./src/pages/Home.js", import.meta.url),
-        middleware: (_request: Request, params: HandlerParams) => params["server_time"] = `${Date.now()}`,
-        render: (_request: Request, params: HandlerParams) => {
-          const appHTML = renderToString(Home(params), null, null)
-          return htmlTemplate({
-            appHTML,
-            title: `<title>Peko</title>`,
-            modulepreload: `<script modulepreload="true" type="module" src="/pages/Home.js"></script>`,
-            hydrationScript: `<script type="module">
-                import { hydrate } from "https://npm.reversehttp.com/preact,preact/hooks,htm/preact,preact-render-to-string";
-                import Home from "/pages/Home.js";
-                hydrate(Home(${JSON.stringify(params)}), document.getElementById("root"))
-            </script>`
-          })
-        },
-        cacheLifetime: 6000 // <- even with a specified cacheLifetime this page will never change because it's params are different in each request
-    },
-    {
-        route: "/about",
-        srcURL: new URL("./src/pages/About.js", import.meta.url),
-        render: () => {
-          const appHTML = renderToString(About(), null, null)
-          return  htmlTemplate({
-            appHTML,
-            title: `<title>Peko | About</title>`,
-            modulepreload: `<script modulepreload="true" type="module" src="/pages/About.js"></script>`,
-            hydrationScript: `<script type="module">
-                import { hydrate } from "https://npm.reversehttp.com/preact,preact/hooks,htm/preact,preact-render-to-string";
-                import About from "/pages/About.js";
-                hydrate(About(), document.getElementById("root"))
-            </script>`
-          })
-        }
-        // cacheLifetime: 6000 <- this can be omitted as page content doesn't change and cacher will default to a lifetime of Infinity
-    }
-]
+
 ssrRoutes.forEach(ssrRoute => Peko.addSSRRoute(ssrRoute))
 
 // Static src routes for loading into client
